# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Scientific Publication Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /papers:
    get:
      operationId: listPublishedPapers
      tags: [papers]
      summary: List all published papers (public)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Paginated published papers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperListResponse"
  /papers/{id}:
    get:
      operationId: getPaper
      tags: [papers]
      summary: Get a paper by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paper details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Paper stats (wow endpoint) ───────────────────────────────────────────────
...
  /my/papers:
    get:
      operationId: listMyPapers
      tags: [papers]
      summary: List current author's submissions
      responses:
        "200":
          description: Author's papers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createPaper
      tags: [papers]
      summary: Create a new paper draft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperInput"
      responses:
        "201":
          description: Draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /my/papers/{id}:
    get:
      operationId: getMyPaper
      tags: [papers]
      summary: Get author's own paper with full detail (history, AI report)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full paper detail including history and AI report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updatePaper
      tags: [papers]
      summary: Update a draft or revision paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaperUpdate"
      responses:
        "200":
          description: Updated paper
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
  /my/papers/{id}/submit:
    post:
      operationId: submitPaper
      tags: [papers]
      summary: Submit a draft or revised paper for review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paper submitted — AI review triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
        "400":
          description: Cannot submit paper in current state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Verifier queue ───────────────────────────────────────────────────────────
  /verifier/queue:
    get:
      operationId: getVerifierQueue
      tags: [reviews]
      summary: Get papers in verifier's review queue
      responses:
        "200":
          description: Papers assigned to this verifier
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaperSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /verifier/papers/{id}/review:
    post:
      operationId: submitReview
      tags: [reviews]
      summary: Submit a review decision on a paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Admin ────────────────────────────────────────────────────────────────────
...
  /users/{id}:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get public user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [USER, VERIFIER, ADMIN]
        verifierLayer:
          type: ["integer", "null"]
        affiliation:
          type: ["string", "null"]
        orcidId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    PaperInput:
      type: object
      required: [title, abstract, keywords, category, pdfUrl, content]
      properties:
        title:
          type: string
        abstract:
          type: string
        keywords:
          type: array
          items:
            type: string
        category:
          type: string
        pdfUrl:
          type: string
        content:
          type: string
        coAuthors:
          type: array
          items:
            type: string
    PaperUpdate:
      type: object
      properties:
        title:
          type: string
        abstract:
          type: string
        keywords:
          type: array
          items:
            type: string
        category:
          type: string
        pdfUrl:
          type: string
        content:
          type: string
        coAuthors:
          type: array
          items:
            type: string
    PaperSummary:
      type: object
      required: [id, title, status, category, authorId, authorName, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        abstract:
          type: string
        keywords:
          type: array
          items:
            type: string
        category:
          type: string
        status:
          type: string
          enum: [DRAFT, SUBMITTED, AI_REVIEW, AI_PASSED, AI_FAILED, LAYER_2_REVIEW, LAYER_2_APPROVED, LAYER_3_REVIEW, LAYER_3_APPROVED, PUBLISHED, REVISION_REQUESTED, REJECTED]
        currentLayer:
          type: integer
        doi:
          type: ["string", "null"]
        authorId:
          type: integer
        authorName:
          type: string
        pdfUrl:
          type: string
        publishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaperDetail:
      allOf:
        - $ref: "#/components/schemas/PaperSummary"
        - type: object
          properties:
            aiReport:
              $ref: "#/components/schemas/AiReport"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/StatusHistoryEntry"
            assignedVerifierId:
              type: ["integer", "null"]
            assignedVerifierName:
              type: ["string", "null"]
            coAuthors:
              type: array
              items:
                type: string
    PaperListResponse:
      type: object
      required: [papers, total, page, limit]
      properties:
        papers:
          type: array
          items:
            $ref: "#/components/schemas/PaperSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── AI Report ─────────────────────────────────────────────────────────────
    AiReport:
      type: object
      required: [score, passed]
      properties:
        score:
          type: number
        passed:
          type: boolean
        wordCount:
          type: integer
        aiGeneratedLikelihood:
          type: number
        plagiarismScore:
          type: number
        structureComplete:
          type: boolean
        citationValid:
          type: boolean
        grammarScore:
          type: number
        issues:
          type: array
          items:
            $ref: "#/components/schemas/AiIssue"
    AiIssue:
      type: object
      required: [type, severity, message]
      properties:
        type:
          type: string
        severity:
          type: string
          enum: [critical, warning, info]
        message:
          type: string
        location:
          type: ["string", "null"]

    # ─── Review ────────────────────────────────────────────────────────────────
    ReviewInput:
      type: object
      required: [decision, comments]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED, REVISION]
        comments:
          type: string
        inlineNotes:
          type: array
          items:
            $ref: "#/components/schemas/InlineNote"
    Review:
      type: object
      required: [id, paperId, reviewerId, reviewerName, layer, decision, comments, createdAt]
      properties:
        id:
          type: integer
        paperId:
          type: integer
        reviewerId:
          type: integer
        reviewerName:
          type: string
        layer:
          type: integer
        decision:
          type: string
          enum: [APPROVED, REJECTED, REVISION]
        comments:
          type: string
        inlineNotes:
          type: array
          items:
            $ref: "#/components/schemas/InlineNote"
        createdAt:
          type: string
          format: date-time
    InlineNote:
      type: object
      required: [note]
      properties:
        page:
          type: integer
        position:
          type: string
        note:
          type: string

    # ─── StatusHistory ─────────────────────────────────────────────────────────
    StatusHistoryEntry:
      type: object
      required: [id, fromStatus, toStatus, changedBy, createdAt]
      properties:
        id:
          type: integer
        fromStatus:
          type: ["string", "null"]
        toStatus:
          type: string
        changedBy:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── Admin ops ─────────────────────────────────────────────────────────────
...